requirements.txt
APScheduler==3.6.3
backoff==1.11.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
six==1.16.0
termcolor==1.1.0
tornado==6.1
tzlocal==3.0
urllib3==1.26.6
